HystrixThreadPoolAdapter supports hippo4j config
